+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 - Add strength and angle sensor for Xiaomi DJT11LM
 - Add smoke density, self test and sensivity support for Xiaomi JTYJ-GD-01LM/BW
+- Ability to remove device from zigbee network using plugin UI
 
 New devices:
 - iCasa Zigbee AC dimmer